'SPELL CORRECTOR GUI USING TKINTER'
This GUI appliaction is going to create an main window with the help of Tkinter library which is most commanly used methods. Python with Tlkinter is the fastest and easiest way to create any GUI application also Tkinter include many functions and methods in it some of the methods are used in building this project when u come across the project. where the used to correct the incorrect word entered by the user. It is an simple GUI applilaction window for an user to understand what actually the program does!.